fix: rename useGetCountries to useCountries
